requirements.txt
aiocfscrape==1.0.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-proxy==0.1.2
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
better-proxy==1.1.5
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
cloudscraper==1.2.71
colorama==0.4.6
DateTime==5.5
fake-useragent==1.5.1
Faker==29.0.0
frozenlist==1.4.1
furl==2.1.3
idna==3.10
loguru==0.7.2
multidict==6.0.5
orderedmultidict==1.0.1
pillow==11.0.0
propcache==0.2.0
pyaes==1.6.1
pycryptodome==3.20.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyjsparser==2.7.1
pyparsing==3.1.4
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
soupsieve==2.5
TgCrypto==1.2.5
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
websockets==12.0
win32-setctime==1.1.0
yarl==1.17.1
zope.interface==6.4.post2
